Fix fp16 NaN: use mixed precision (fp32 master + autocast + GradScaler)
--fp16 previously loaded MASTER weights in fp16, so AdamW's second-moment state
underflowed and the loss went NaN within ~10 steps. Now master weights are
always fp32 and --fp16 enables autocast(fp16) compute + a GradScaler for the
backward/step — the stable, standard way to get fp16 speed/memory on a T4.

- distill.py, distill_v2.py: fp32 weights, autocast forward, scaler.scale/
  unscale_/step/update around grad accumulation.
- evaluate.py: load teacher + student in fp32 for gold-standard perplexity
  (size is computed analytically, so load dtype does not change reported GB).

+++ evaluate.py
@@ -151,8 +151,10 @@ def main(args):
         with open(args.results) as f:
             results = json.load(f)
 
+    # fp32 for stable, gold-standard perplexity (size is computed analytically,
+    # so load dtype does not affect the reported GB).
     teacher = AutoModelForCausalLM.from_pretrained(
-        args.model_name, torch_dtype=torch.float16).to(device)
+        args.model_name, torch_dtype=torch.float32).to(device)
     config = teacher.config
 
     # ---- baselines (teacher + 4-bit): compute once, reuse across ranks ----
@@ -184,7 +186,7 @@ def main(args):
         rank = ckpt.get("rank", args.rank)
         student = ConsolidatedQwen(args.model_name, rank=rank,
                                    alpha=ckpt.get("alpha", 16.0),
-                                   dtype=torch.float16).to(device)
+                                   dtype=torch.float32).to(device)
         student.load_consolidated_state_dict(ckpt["state_dict"])
         params, gb, breakdown = student_size(teacher, student, config)
         ppl = perplexity(student, blocks, device, args.eval_batch_size)
